+++ app/src/main/java/com/steganoapp/steganography/SteganoMethod.java
@@ -3,7 +3,7 @@ package com.steganoapp.steganography;
 import org.opencv.core.Mat;
 
 public interface SteganoMethod {
-    Mat encode(Mat picture, byte[] message);
+    Mat encode(Mat picture, int[] message);
     byte[] decode(Mat picture);
 
     static SteganoMethod getInstance(String methodName) {
